ai-agent/requirements.txt
videosdk-agents>=0.1.0
videosdk-plugins-google>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
